CAMEL-11792: File/ftp producer should not do file exists check if the fileExists is TryRename as it state in the docs it does not do any exists check. The FTP producer will then not create a new connection when uploading files to the ftp server.

+++ camel-core/src/main/java/org/apache/camel/component/file/GenericFileProducer.java
@@ -154,7 +154,7 @@ public class GenericFileProducer<T> extends DefaultProducer {
                 }
 
                 // delete any pre existing temp file
-                if (operations.existsFile(tempTarget)) {
+                if (endpoint.getFileExist() != GenericFileExist.TryRename && operations.existsFile(tempTarget)) {
                     log.trace("Deleting existing temp file: {}", tempTarget);
                     if (!operations.deleteFile(tempTarget)) {
                         throw new GenericFileOperationFailedException("Cannot delete file: " + tempTarget);
